Improve CrawlerFactory
the crawler factory is not static anymore.
It has a register method to add new CrawlerTypes.
They are already split between interfaces and
"real" types. Interface Crawlers are not yet handled.

The crawler factory find the most appropriate
crawler type by inspecting the inheritance chain
of the object. System.Object is always at the end
of such a chain. The object crawler is preinstalled
in any crawler factory.

Any crawler is assigned to a factory, that is used
when the crawler follows a path in order to find
the next crawler.

The Path of a Crawler is now base on a
SingleLinkedList<CrawlerItem>
instead of SingleLinkedList<Crawler>