added input crawler
sometime user input is necessary. The input crawler is used for this.
In the datamodel you can ask for input by just making a property of type
Ananse.Input
the appropriate crawler will be the input crawler.
The presenter will provide mean to input text if it comes along an
input crawler.